refactor : 미팅 결과에 의한 코드 리팩토링
- 요청/응답에 대한 클래스명을 Request/Response로 통일
- 각각 request/response 패키지로 위치 변경
- 패키지 위치 변경으로 인한 리팩토링
- enum을 소문자로 입력한 경우의 처리를 위한 코드 리팩토링
