+++ src/main/java/com/zerototen/savegame/repository/RecordRepositoryImpl.java
@@ -3,9 +3,9 @@ package com.zerototen.savegame.repository;
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.zerototen.savegame.domain.QRecord;
-import com.zerototen.savegame.domain.entity.Record;
 import com.zerototen.savegame.domain.dto.RecordAnalysisServiceDto;
+import com.zerototen.savegame.domain.entity.QRecord;
+import com.zerototen.savegame.domain.entity.Record;
 import com.zerototen.savegame.domain.type.Category;
 import java.time.LocalDate;
 import java.util.List;
